+++ templates/web-angular/lib/app_component.dart
@@ -1,7 +1,7 @@
 // Copyright (c) __year__, __author__. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
-import 'package:angular2/angular2.dart';
+import 'package:angular/angular.dart';
 import 'package:angular_components/angular_components.dart';
 
 import 'src/todo_list/todo_list_component.dart';