+++ templates/web-angular/lib/src/todo_list/todo_list_component.dart
@@ -3,7 +3,7 @@
 
 import 'dart:async';
 
-import 'package:angular2/angular2.dart';
+import 'package:angular/angular.dart';
 import 'package:angular_components/angular_components.dart';
 
 import 'todo_list_service.dart';