+++ templates/web-angular/lib/src/todo_list/todo_list_service.dart
@@ -3,7 +3,7 @@
 
 import 'dart:async';
 
-import 'package:angular2/core.dart';
+import 'package:angular/core.dart';
 
 /// Mock service emulating access to a to-do list stored on a server.
 @Injectable()