+++ templates/web-angular/test/app_test.dart
@@ -2,7 +2,7 @@
 @TestOn('browser')
 import 'dart:async';
 
-import 'package:angular2/angular2.dart';
+import 'package:angular/angular.dart';
 import 'package:angular_test/angular_test.dart';
 import 'package:pageloader/objects.dart';
 import 'package:test/test.dart';